feature(vscode-iframe) use sample code with shimmer
